pyproject.toml
[project]
name = "icbhi-ast-sam"
version = "0.1.0"
description = "Respiratory sound classification on ICBHI 2017 with AST + SAM"
requires-python = ">=3.12,<3.13"
dependencies = [
    "librosa>=0.11.0",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "soundfile>=0.13.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "tqdm>=4.67.3",
    "transformers>=5.6.2",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchaudio = [{ index = "pytorch-cu124" }]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "kaggle>=2.1.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.12",
]
